--- client/src/constants/Data/additionalLinks.js
@@ -1,8 +1,0 @@
-const additionalLinks = [
-  { label: "Twitter", url: "https://twitter.com/kannanmoudgalya" },
-  { label: "Tedx Talk", url: "https://www.youtube.com/watch?v=JaX_uD8JFpI" },
-  { label: "About PI", url: "" },
-  { label: "Partnered with Nasscom", url: "https://futureskills.nasscom.in/" },
-];
-
-export { additionalLinks };
