--- client/src/constants/Data/announcements.js
@@ -1,7 +1,0 @@
-const announcements = [
-  "Spoken Tutorials in Korku language helps lessen Child malnutrition in Melghat.",
-  "Spoken Tutorials for Education, Employment and Empowerment | Kannan Moudgalya | TEDxNMIMSShirpur.",
-  "Spoken Tutorials for Education, Employment and Empowerment",
-];
-
-export { announcements };
