--- client/src/constants/Data/awards.js
@@ -1,8 +1,0 @@
-const awards = [
-  "Re-Engineering India Award 2020",
-  "Reimagine Education Award 2015",
-  "Nomination for HT for Mumbai Awards 2015",
-  "Google MOOCs Research Award 2015",
-];
-
-export { awards };
