--- client/src/constants/Data/communityInsights.js
@@ -1,22 +1,0 @@
-const coumminityInsights = [
-  {
-    title: "Forum",
-    description:
-      "Spoken Tutorial Forums is a friendly online discussion forum. You can join existing discussions or start new topics, and get lots of replies from the Spoken Tutorial community. Registration to Forums is completely free and takes only one minute.",
-    url: "#",
-  },
-  {
-    title: "News / Events",
-    description:
-      "Check out the amazing testimonials, media articles, events, collaboration, seminars, talks and campaigns about the Spoken Tutorial Project. Browse through our gallery and see what the Project has been doing across India.",
-    url: "#",
-  },
-  {
-    title: "Software Training",
-    description:
-      "The Spoken Tutorial Project is about teaching and learning a particular FOSS (Free and Open Source Software) like Linux, Scilab, LaTeX, PHP & MySQL, Java, C/C++, LibreOffice etc. via an easy Video tool - Spoken Tutorials.",
-    url: "#",
-  },
-];
-
-export { coumminityInsights };
